move float4 stuff to separate math file, created utils for shared actions
